Created whole new section Client Overview to get a global view of all your clients things, merged domains and assets into one
