+++ includes/inc_client_overview_all.php
@@ -1,0 +1,12 @@
+<?php
+
+require_once "config.php";
+require_once "functions.php";
+require_once "check_login.php";
+require_once "page_title.php";
+require_once "header.php";
+require_once "top_nav.php";
+require_once "client_overview_side_nav.php";
+require_once "inc_wrapper.php";
+require_once "inc_alert_feedback.php";
+require_once "filter_header.php";
