+++ modals/domain_add_modal.php
@@ -8,7 +8,6 @@
                 </button>
             </div>
             <form action="post.php" method="post" autocomplete="off">
-                <input type="hidden" name="client_id" value="<?php echo $client_id; ?>">
                 <div class="modal-body bg-white">
 
                     <ul class="nav nav-pills nav-justified mb-3">
@@ -46,6 +45,38 @@
                                     <input type="text" class="form-control" name="description" placeholder="Short Description">
                                 </div>
                             </div>
+
+                            <?php if ($client_url) { ?>
+                                <input type="hidden" name="client_id" value="<?php echo $client_id; ?>">
+                            <?php } else { ?>
+
+                                <div class="form-group">
+                                    <label>Client <strong class="text-danger">*</strong> / <span class="text-secondary">Use Primary Contact</label>
+                                    <div class="input-group">
+                                        <div class="input-group-prepend">
+                                            <span class="input-group-text"><i class="fa fa-fw fa-user"></i></span>
+                                        </div>
+                                        <select class="form-control select2" name="client_id" required>
+                                            <option value="">- Select Client -</option>
+                                            <?php
+
+                                            $sql = mysqli_query($mysqli, "SELECT client_id, client_name FROM clients WHERE client_archived_at IS NULL $access_permission_query ORDER BY client_name ASC");
+                                            while ($row = mysqli_fetch_array($sql)) {
+                                                $client_id = intval($row['client_id']);
+                                                $client_name = nullable_htmlentities($row['client_name']); ?>
+                                                <option value="<?php echo $client_id; ?>"><?php echo $client_name; ?></option>
+
+                                            <?php } ?>
+                                        </select>
+                                        <div class="input-group-append">
+                                            <div class="input-group-text">
+                                                <input type="checkbox" name="use_primary_contact" value="1">
+                                            </div>
+                                        </div>
+                                    </div>
+                                </div>
+
+                            <?php } ?>
 
                             <div class="form-group">
                                 <label>Registrar</label>
